PC2/mongodb_to_mysql/requirements.txt
paho-mqtt==2.1.0 
mysql-connector-python==9.3.0
